This is a very large commit.
Decision tree models targeting various cuts of the grades at various samples of the available data are created.  The scripts are saved; the models are saved; and the training/testing data is located in the "Data" folder (ignored in Git.)

Decision Tree vD0 uses complete cases of all of the available data.

Notes from Friday's meeting and musings are saved.