Changed courseScatter by adding libraries, reducing the radius of the ellipse to 0.5 IQR, and added an ability to exactly plot the 25th and 75th percentile with rectangles (needs attention.)
Also modified the default colors as the dark blue was too close to the dark green.

Also added an ability to return the input data with the cluster assignment.

Started a "Course Guide" which I'm envisioning as first pass is an assignment to one of three clusters of courses, second pass is diagrams of the courses, and fourth is a table of the quantiles per course (very specific).

In IQR Guidelines I play and iterate with various concepts until arriving at the plan above.